ref(constants): split consts to leaves to break import cycles
Relocate symbols that lived in the `constants/index` god-barrel and a few
type hubs into pure leaf modules, so `constants/index` only imports leaves
and drops out of the frontend type-import strongly-connected component.

- `constants/scopes`: `API_ACCESS_SCOPES`, `ALLOWED_SCOPES`, and `Scope`
- `types/dataCategory`: `DataCategory`, `DataCategoryExact`, `DataCategoryInfo`
- `types/permissions`: `PermissionValue`, `Permissions`, `PermissionResource`
- `types/roles`: `BaseRole`, `OrgRole`, `TeamRole`

The original hubs re-export the moved symbols so existing consumers are
untouched. `DataCategoryInfo.getProductLink` is narrowed to accept
`{slug: string}` so the leaf needs no `Organization` import.
