+++ static/app/constants/scopes.tsx
@@ -1,0 +1,52 @@
+export const API_ACCESS_SCOPES = [
+  'alerts:read',
+  'alerts:write',
+  'event:admin',
+  'event:read',
+  'event:write',
+  'member:admin',
+  'member:read',
+  'member:write',
+  'org:admin',
+  'org:ci',
+  'org:integrations',
+  'org:read',
+  'org:write',
+  'project:admin',
+  'project:distribution',
+  'project:read',
+  'project:releases',
+  'project:write',
+  'team:admin',
+  'team:read',
+  'team:write',
+] as const;
+
+export const ALLOWED_SCOPES = [
+  'alerts:read',
+  'alerts:write',
+  'event:admin',
+  'event:read',
+  'event:write',
+  'member:admin',
+  'member:invite',
+  'member:read',
+  'member:write',
+  'org:admin',
+  'org:billing',
+  'org:ci',
+  'org:integrations',
+  'org:read',
+  'org:superuser', // not an assignable API access scope
+  'org:write',
+  'project:admin',
+  'project:distribution',
+  'project:read',
+  'project:releases',
+  'project:write',
+  'team:admin',
+  'team:read',
+  'team:write',
+] as const;
+
+export type Scope = (typeof ALLOWED_SCOPES)[number];
