+++ static/app/types/permissions.tsx
@@ -1,0 +1,14 @@
+export type PermissionValue = 'no-access' | 'read' | 'write' | 'admin';
+
+export type Permissions = {
+  Event: PermissionValue;
+  Member: PermissionValue;
+  Organization: PermissionValue;
+  Project: PermissionValue;
+  Release: PermissionValue;
+  Team: PermissionValue;
+  Alerts?: PermissionValue;
+  Distribution?: PermissionValue;
+};
+
+export type PermissionResource = keyof Permissions;
